Move TaskDependencyGraphConfiguration to `graph-publisher`
