+++ library/plugins/talaiot-legacy/src/main/kotlin/com/cdsap/talaiot/plugin/TalaiotPublisherConfiguration.kt
@@ -4,8 +4,7 @@ import com.cdsap.talaiot.configuration.*
 import com.cdsap.talaiot.publisher.OutputPublisherConfiguration
 import com.cdsap.talaiot.publisher.Publisher
 import com.cdsap.talaiot.publisher.hybrid.HybridPublisherConfiguration
-
-
+import com.cdsap.talaiot.publisher.graph.TaskDependencyGraphConfiguration
 import groovy.lang.Closure
 import org.gradle.api.Project
